requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
line-bot-sdk==1.17.0
mccabe==0.6.1
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.11
wrapt==1.12.1
